test: enable acceptance tests on emulator (#70)
* test: add mock server and tests

Adds a mock server and several end-to-end tests using the mock server.
Also fixes:
- Invalid function calls in migrations
- Unnecessary use of snapshots for INFORMATION_SCHEMA queries
- Unnecessary SELECT 1 query each time a connection is checked out of the pool

* chore: fix linting issues

* fix: wait until server is running before proceding

* test: enable acceptance tests on emulator

Enables running the acceptance integration tests on the emulator. These finish in ~20 seconds
instead of after several hours. All tests are green, except for one that is skipped, as
Cloud Spanner does not support changing a nullable column to a not-null column for a table
that already exists.

Updates #65

* test: enable acceptance tests on emulator

Enables running the acceptance integration tests on the emulator. These finish in ~20 seconds
instead of after several hours. All tests are green, except for one that is skipped, as
Cloud Spanner does not support changing a nullable column to a not-null column for a table
that already exists.

Updates #65

* chore: fix linting issues

* fix: reset primary key prefix after change definition

* test: run acceptance tests on emulator

* fix: set correct task name

* build: run on all supported ruby versions

* fix: check instance value + document join table